FRONTEND: Add logged in/out variants of header bar.
Implements a logged in and logged out variant of the header bar. Logged
out variant of the header bar shows a button with the text 'LOGIN' to
prompt users to log in. Logged in variant replaces the LOGIN button with
an avatar icon of the user. Clicking on the avatar opens up a floating
menu where users can choose further actions.

Implementations are done in frontend only, actual functionality of these
buttons are yet to be added.