Revert "Resolve conflicts"
This reverts commit 97422f35a74676a14cda01044d8049656708ad1b.

Resolve conflicts again with file naming.

Strange naming conflict occuring with the directory name located at
src/userstorage.

On my local PC, the folder is in PascalCase: UserStorage
On github, the folder is in lowercase: userstorage

Pulling and pushing did not change the directory name, hence the
confusion. Somehow pulled the PascalCase version, yet the remote repo on
github is in lowercase, hence the reason for commit #97422f3

Since vercel deploys from the github repo, I will change the directory
name on my PC to lowercase, and revert the changes made in #97422f3.