Expand onClick functionality of ClickableCard.
Previously, ClickableCard only accepts a `URL` prop that it will
use for redirection as part of its onClick behaviour. This
restricts the functionality of ClickableCard to only perform
redirection when clicked.

Now ClickableCard accepts a generic `onClick` prop that takes
in any function. If `onClick` is specified, ClickableCard will
call the function passed to `onClick` and ignore the `URL` prop.